# src/utils/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Survey API Overview",
    "description": "Survey API documentation generated by swagger-autogen",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/surveys/statistics": {
      "get": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to get survey statistics",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/surveys/{id}": {
      "get": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to get a survey by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Survey ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to update a survey by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Survey ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "survey",
            "in": "body",
            "description": "Survey details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "required": {
                  "type": "array",
                  "example": [
                    "title",
                    "createdBy",
                    "questions",
                    "expirationDate"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Updated Survey Title"
                        }
                      }
                    },
                    "description": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Updated Survey Description"
                        }
                      }
                    },
                    "questions": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "questionText": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "What is your age?"
                                    }
                                  }
                                },
                                "type": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "enum": {
                                      "type": "array",
                                      "example": [
                                        "textInput",
                                        "numberInput",
                                        "checkbox",
                                        "toggle",
                                        "multipleChoice",
                                        "radio",
                                        "stars"
                                      ],
                                      "items": {
                                        "type": "string"
                                      }
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "numberInput"
                                    }
                                  }
                                },
                                "options": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "array"
                                    },
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "type": {
                                          "type": "string",
                                          "example": "string"
                                        },
                                        "example": {
                                          "type": "string",
                                          "example": "Option A"
                                        }
                                      }
                                    }
                                  }
                                },
                                "order": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "required": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "boolean"
                                    },
                                    "example": {
                                      "type": "boolean",
                                      "example": true
                                    }
                                  }
                                }
                              }
                            },
                            "example": {
                              "type": "object",
                              "properties": {
                                "questionText": {
                                  "type": "string",
                                  "example": "What is your age?"
                                },
                                "type": {
                                  "type": "string",
                                  "example": "numberInput"
                                },
                                "options": {
                                  "type": "array",
                                  "example": [],
                                  "items": {}
                                },
                                "order": {
                                  "type": "number",
                                  "example": 1
                                },
                                "required": {
                                  "type": "boolean",
                                  "example": true
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "expirationDate": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "date-time"
                        },
                        "example": {
                          "type": "string",
                          "example": "2023-12-31T23:59:59Z"
                        }
                      }
                    },
                    "createdBy": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109ca"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Updated Survey Title"
                    },
                    "description": {
                      "type": "string",
                      "example": "Updated Survey Description"
                    },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionText": {
                            "type": "string",
                            "example": "What is your favorite hobby?"
                          },
                          "type": {
                            "type": "string",
                            "example": "textInput"
                          },
                          "options": {
                            "type": "array",
                            "example": [],
                            "items": {}
                          },
                          "order": {
                            "type": "number",
                            "example": 2
                          },
                          "required": {
                            "type": "boolean",
                            "example": false
                          }
                        }
                      }
                    },
                    "expirationDate": {
                      "type": "string",
                      "example": "2023-12-31T23:59:59Z"
                    },
                    "createdBy": {
                      "type": "string",
                      "example": "60d0fe4f5311236168a109ca"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to delete a survey by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Survey ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/surveys/{id}/public": {
      "get": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to get a public survey by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Public Survey ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/surveys/": {
      "get": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to list surveys",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit of surveys per page",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to create a new survey",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "survey",
            "in": "body",
            "description": "Survey details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "required": {
                  "type": "array",
                  "example": [
                    "title",
                    "createdBy",
                    "questions",
                    "expirationDate"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Survey Title"
                        }
                      }
                    },
                    "description": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Survey Description"
                        }
                      }
                    },
                    "questions": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "questionText": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "What is your name?"
                                    }
                                  }
                                },
                                "type": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "enum": {
                                      "type": "array",
                                      "example": [
                                        "textInput",
                                        "numberInput",
                                        "checkbox",
                                        "toggle",
                                        "multipleChoice",
                                        "radio",
                                        "stars"
                                      ],
                                      "items": {
                                        "type": "string"
                                      }
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "textInput"
                                    }
                                  }
                                },
                                "options": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "array"
                                    },
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "type": {
                                          "type": "string",
                                          "example": "string"
                                        },
                                        "example": {
                                          "type": "string",
                                          "example": "Option 1"
                                        }
                                      }
                                    }
                                  }
                                },
                                "order": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "number"
                                    },
                                    "example": {
                                      "type": "number",
                                      "example": 1
                                    }
                                  }
                                },
                                "required": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "boolean"
                                    },
                                    "example": {
                                      "type": "boolean",
                                      "example": true
                                    }
                                  }
                                }
                              }
                            },
                            "example": {
                              "type": "object",
                              "properties": {
                                "questionText": {
                                  "type": "string",
                                  "example": "What is your name?"
                                },
                                "type": {
                                  "type": "string",
                                  "example": "textInput"
                                },
                                "options": {
                                  "type": "array",
                                  "example": [],
                                  "items": {}
                                },
                                "order": {
                                  "type": "number",
                                  "example": 1
                                },
                                "required": {
                                  "type": "boolean",
                                  "example": true
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "expirationDate": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "date-time"
                        },
                        "example": {
                          "type": "string",
                          "example": "2023-12-31T23:59:59Z"
                        }
                      }
                    },
                    "createdBy": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109ca"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Survey Title"
                    },
                    "description": {
                      "type": "string",
                      "example": "Survey Description"
                    },
                    "questions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "questionText": {
                            "type": "string",
                            "example": "What are your favorite colors?"
                          },
                          "type": {
                            "type": "string",
                            "example": "checkbox"
                          },
                          "options": {
                            "type": "array",
                            "example": [
                              "Red",
                              "Green",
                              "Blue"
                            ],
                            "items": {
                              "type": "string"
                            }
                          },
                          "order": {
                            "type": "number",
                            "example": 3
                          },
                          "required": {
                            "type": "boolean",
                            "example": true
                          }
                        }
                      }
                    },
                    "expirationDate": {
                      "type": "string",
                      "example": "2023-12-31T23:59:59Z"
                    },
                    "createdBy": {
                      "type": "string",
                      "example": "60d0fe4f5311236168a109ca"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/surveys/{id}/responses": {
      "post": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to save a survey response",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Survey ID"
          },
          {
            "name": "response",
            "in": "body",
            "description": "Survey response details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "responses": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "questionId": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "60d0fe4f5311236168a109cd"
                                    }
                                  }
                                },
                                "answer": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "Sample Answer"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/surveys/{id}/result": {
      "get": {
        "tags": [
          "Survey"
        ],
        "description": "Endpoint to get the result of a survey by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Survey ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/get/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/list": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notifications/list": {
      "get": {
        "tags": [
          "Notification"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}